[plm_auto_internalref] :fire: rename models' folder.
Also, fixed some `pep8` issues.
