+++ plm_auto_internalref/models/__init__.py
@@ -1,4 +1,3 @@
-# -*- encoding: utf-8 -*-
 ##############################################################################
 #
 #    OmniaSolutions, Your own solutions
@@ -20,29 +19,10 @@
 #
 ##############################################################################
 
-'''
+"""
 Created on 9 Dec 2016
 
 @author: Daniel Smerghetto
-'''
+"""
 
-from odoo import models
-from odoo import api
-import logging
-
-
-class ProductProductExtension(models.Model):
-    _name = 'product.product'
-    _inherit = 'product.product'
-
-    @api.model
-    def create(self, vals):
-        engineering_code = vals.get('engineering_code', '')
-        engineering_revision = vals.get('engineering_revision', 0)
-        if (engineering_code and not vals.get('default_code')) or (engineering_code and engineering_revision > 0):
-            vals['default_code'] = '%s_%s' % (engineering_code, engineering_revision)
-            logging.info('Internal ref set value %s on engineering_code: %r' % (vals['default_code'], engineering_code))
-        return super(ProductProductExtension, self).create(vals)
-
-
-# vim:expandtab:smartindent:tabstop=4:softtabstop=4:shiftwidth=4:
+from . import product_product
